Delete resource anyway when work owned by multiple hubs is deleting
If the resource applied by work has multiple owners, the resource will
not be removed when the work is deleted. However, we need to remove the
resource anyway if the resource is owned by different hubs. This is
because the hub could be switched to a new one, while previous hub
cannot own the resource anymore.

Signed-off-by: Jian Qiu <jqiu@redhat.com>